# docs/openapi/baseline.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SpecProof Phase 1",
    "description": "Reliable Verification Job API",
    "version": "0.1.0"
  },
  "paths": {
    "/jobs": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Create Job",
        "description": "Create a verification job and its Outbox event in ONE MySQL transaction.\n\nThe job row and the outbox row commit together: if the API crashes before\nresponding, the relay still publishes the event (at-least-once). The\nconsumer is idempotent, so a duplicate delivery cannot double-process.\n\nReturns 202 with job_id. A real 503 when MySQL is unreachable — the API\nnever pretends a job was accepted when nothing was persisted.",
        "operationId": "create_job_jobs_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Job Jobs Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Jobs",
        "description": "List the most recent verification jobs.",
        "operationId": "list_jobs_jobs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Jobs Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job",
        "description": "Cancel a QUEUED/RUNNING job (CAS; terminal jobs are immutable).",
        "operationId": "cancel_job_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Cancel Job Jobs  Job Id  Cancel Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "description": "Return one job's status from MySQL (the business source of truth).",
        "operationId": "get_job_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Job Jobs  Job Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/summary": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job Summary",
        "description": "Return the persisted pipeline summary (matrix/findings/capsules).",
        "operationId": "get_job_summary_jobs__job_id__summary_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Job Summary Jobs  Job Id  Summary Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/progress": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Job Progress",
        "description": "SSE endpoint for job progress. Supports reconnection via Last-Event-ID.\n\nEvent format:\n    id: <stream-entry-id>\n    event: progress\n    data: {\"node\": \"...\", \"status\": \"...\", \"percent\": N, \"message\": \"...\"}\n\nThe client disconnects, then reconnects with Last-Event-ID header set to\nthe last received event id. The server resumes from that position.",
        "operationId": "job_progress_jobs__job_id__progress_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreateRequest": {
        "properties": {
          "repo_path": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Repo Path"
          },
          "base_ref": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Base Ref"
          },
          "head_ref": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Head Ref"
          },
          "spec_path": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Spec Path"
          },
          "depth": {
            "type": "string",
            "pattern": "^(FAST)$",
            "title": "Depth",
            "default": "FAST"
          }
        },
        "type": "object",
        "required": [
          "repo_path",
          "base_ref",
          "head_ref",
          "spec_path"
        ],
        "title": "JobCreateRequest",
        "description": "Job submission payload. Paths/refs are validated server-side."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}